fix(backend): 对齐实体映射与 Flyway 迁移，修复全新 MySQL 库建库即失败
迁移脚本此前用了 H2 的 CLOB 类型与 MySQL 保留字 condition，全新库跑到 V5 就中断；同时实体与脚本在 @Lob 列类型、枚举映射、列名上从未对齐过，ddl-auto=validate 必然失败。本次逐项对齐后，全新 MySQL 库可完整跑通 27 个迁移并通过结构校验。

要点：@Lob String 按 length 推导类型（默认 255 会落成 tinytext，装不下堆栈与权限 JSON）；枚举统一走 @Enumerated(STRING) 并改为原生 ENUM 列；t_detection_event 的 evidence_* 前缀属笔误，按实体字段名改正；AlertRule.condition 改名 rule_condition（该保留字此前让 Hibernate 生成 DDL 直接报错，线上库根本没建出这张表）。

+++ ptv-backend/src/main/resources/db/migration/V13__v47_threat_intel_ops.sql
@@ -29,7 +29,7 @@ CREATE TABLE IF NOT EXISTS t_ioc_indicator (
     state          VARCHAR(16)  NOT NULL DEFAULT 'OPEN',  -- OPEN / DISARMED / EXPIRED
     subscribed     TINYINT(1)   NOT NULL DEFAULT 0,
     alert_threshold INT NOT NULL DEFAULT 3,
-    hit_count      INT  NOT NULL DEFAULT 0,
+    hit_count      BIGINT NOT NULL DEFAULT 0,
     first_seen     DATETIME(6)  NOT NULL,
     last_seen      DATETIME(6)  NOT NULL,
     PRIMARY KEY (id),
